requirements.txt
python-telegram-bot[job-queue]>=20.3
python-dotenv>=1.0.0
requests>=2.28.0
